CLI and Logging improvments

+++ Addons/Entitas.CodeGenerator.CLI/Program.cs
@@ -21,7 +21,7 @@ namespace Entitas.CodeGenerator.CLI {
             try {
                 switch(args[0]) {
                     case "new":
-                        newConfig();
+                        createNewConfig(args.Any(arg => arg == "-f"));
                         break;
                     case "edit":
                         editConfig();
@@ -59,20 +59,20 @@ namespace Entitas.CodeGenerator.CLI {
 
         static void printUsage() {
             Console.WriteLine(
-@"usage: entitas new      - Creates new Entitas.properties with default values
-       entitas edit     - Opens Entitas.properties
-       entitas doctor   - Checks the config for potential problems
-       entitas diff     - List of unused and invalid data providers, code generators and post processors
-       entitas scan     - Scans and prints available types found in specified assemblies
-       entitas dry      - Simulates generating files without running post processors
-       entitas gen      - Generates files based on Entitas.properties
-       [-v]             - verbose output"
+@"usage: entitas new [-f]     - Creates new Entitas.properties with default values
+       entitas edit         - Opens Entitas.properties
+       entitas doctor       - Checks the config for potential problems
+       entitas diff         - List of unused and invalid data providers, code generators and post processors
+       entitas scan         - Scans and prints available types found in specified assemblies
+       entitas dry          - Simulates generating files without running post processors
+       entitas gen          - Generates files based on Entitas.properties
+       [-v]                 - verbose output"
             );
         }
 
         static void setupLogging(string[] args) {
             if(args.Any(arg => arg == "-v")) {
-                fabl.globalLogLevel = LogLevel.Debug;
+                fabl.globalLogLevel = LogLevel.On;
             } else {
                 fabl.globalLogLevel = LogLevel.Info;
             }
@@ -83,31 +83,40 @@ namespace Entitas.CodeGenerator.CLI {
             });
         }
 
-        static void newConfig() {
-            var types = AppDomain.CurrentDomain.GetAllTypes();
-            var defaultConfig = new CodeGeneratorConfig(
-                new EntitasPreferencesConfig(string.Empty),
-                CodeGeneratorUtil.GetOrderedTypeNames<ICodeGeneratorDataProvider>(types).ToArray(),
-                CodeGeneratorUtil.GetOrderedTypeNames<ICodeGenerator>(types).ToArray(),
-                CodeGeneratorUtil.GetOrderedTypeNames<ICodeGenFilePostProcessor>(types).ToArray()
-            );
-
+        static void createNewConfig(bool force) {
             var currentDir = Directory.GetCurrentDirectory();
-
             var path = currentDir + Path.DirectorySeparatorChar + EntitasPreferences.GetConfigPath();
-            var config = defaultConfig.ToString();
-            File.WriteAllText(path, config);
-            _logger.Info("Created " + path);
-            _logger.Debug(config);
+
+            if(!File.Exists(path) || force) {
+                var types = AppDomain.CurrentDomain.GetAllTypes();
+                var defaultConfig = new CodeGeneratorConfig(
+                    new EntitasPreferencesConfig(string.Empty),
+                    CodeGeneratorUtil.GetOrderedTypeNames<ICodeGeneratorDataProvider>(types).ToArray(),
+                    CodeGeneratorUtil.GetOrderedTypeNames<ICodeGenerator>(types).ToArray(),
+                    CodeGeneratorUtil.GetOrderedTypeNames<ICodeGenFilePostProcessor>(types).ToArray()
+                );
+
+                var config = defaultConfig.ToString();
+                File.WriteAllText(path, config);
+                _logger.Info("Created " + path);
+                _logger.Debug(config);
+
+                editConfig();
+            } else {
+                _logger.Warn(path + " already exists!");
+                _logger.Info("Use entitas new -f to overwrite exiting file.");
+            }
         }
 
         static void editConfig() {
+            _logger.Debug("Opening " + EntitasPreferences.GetConfigPath());
             System.Diagnostics.Process.Start(EntitasPreferences.GetConfigPath());
         }
 
         static void doctor() {
             if(File.Exists(EntitasPreferences.GetConfigPath())) {
                 diff();
+                _logger.Debug("Dry Run");
                 CodeGeneratorUtil
                     .CodeGeneratorFromConfig(EntitasPreferences.GetConfigPath())
                     .DryRun();
@@ -119,7 +128,6 @@ namespace Entitas.CodeGenerator.CLI {
         static void diff() {
             if(File.Exists(EntitasPreferences.GetConfigPath())) {
                 var fileContent = File.ReadAllText(EntitasPreferences.GetConfigPath());
-                _logger.Debug(fileContent);
                 var properties = new Properties(fileContent);
                 var config = new CodeGeneratorConfig(new EntitasPreferencesConfig(fileContent));
                 var types = CodeGeneratorUtil.LoadTypesFromCodeGeneratorAssemblies();
@@ -142,14 +150,14 @@ namespace Entitas.CodeGenerator.CLI {
         static void printUnusedKeys(Properties properties) {
             var unusedKeys = properties.keys.Where(key => !CodeGeneratorConfig.keys.Contains(key));
             foreach(var key in unusedKeys) {
-                _logger.Info("Unused " + key);
+                _logger.Info("Unused key: " + key);
             }
         }
 
         static void printMissingKeys(Properties properties) {
             var missingKeys = CodeGeneratorConfig.keys.Where(key => !properties.HasKey(key));
             foreach(var key in missingKeys) {
-                _logger.Warn("Missing " + key);
+                _logger.Warn("Missing key: " + key);
             }
         }
 
@@ -193,13 +201,13 @@ namespace Entitas.CodeGenerator.CLI {
 
         static void printUnavailable(string[] names) {
             foreach(var name in names) {
-                _logger.Warn("Unavailable " + name);
+                _logger.Warn("Unavailable: " + name);
             }
         }
 
         static void printAvailable(string[] names) {
             foreach(var name in names) {
-                _logger.Info("Available " + name);
+                _logger.Info("Available: " + name);
             }
         }
